requirements.txt
Flask==2.0.2
Pillow==9.0.1
python-dotenv==0.19.2
requests==2.27.1
asyncpg>=0.25.0
py-cord @ git+https://github.com/Pycord-Development/pycord@b183d199cc2b15a761f6e2b5fb4fc886f19b1629
